made the grid spawn logic
it it can spawn up to 4 player starting zones, each color coded, it centers the spawn zones on thier repective axis, you can change the gird size and the size of the playerzonez in the inspecto. What's missing and i don't know if i will have time to implement: what happens if the player zones overlap/if the player zones are bigger the the grid itslef. I will come back to this later